Fixes in the build logic

+++ clear_build.sh
@@ -1,7 +1,20 @@
 #!/bin/bash
 
 ## Clear the python package built files
-rm -rf dist build txm.egg-info
+
+# First check if the package folders [dist build txm.egg-info] exists
+
+if [ -d dist ] && [ -d build ] && [ -d txm.egg-info ]; then
+    rm -rf dist build txm.egg-info
+else
+    echo "No package files to clear"
+fi
 
 ## Clear logs
-rm *.log
+# First check if the log files exists
+
+if [ -f *.log ]; then
+    rm *.log
+else
+    echo "No log files to clear"
+fi
